split up root
