+++ pallets/subtensor/src/coinbase/mod.rs
@@ -2,3 +2,4 @@ use super::*;
 pub mod block_step;
 pub mod root;
 pub mod run_coinbase;
+pub mod block_emission;
