+++ pallets/subtensor/src/coinbase/root.rs
@@ -18,15 +18,11 @@
 use super::*;
 // use crate::epoch::math::*;
 use frame_support::dispatch::Pays;
-use frame_support::storage::{IterableStorageDoubleMap, IterableStorageMap};
+use frame_support::storage::IterableStorageDoubleMap;
 use frame_support::traits::Get;
 use frame_support::weights::Weight;
-use sp_runtime::Saturating;
 use sp_std::vec;
-use substrate_fixed::{
-    transcendental::log2,
-    types::{I64F64, I96F32},
-};
+use substrate_fixed::types::I64F64;
 
 impl<T: Config> Pallet<T> {
     /// Retrieves the unique identifier (UID) for the root network.
@@ -38,37 +34,6 @@ impl<T: Config> Pallet<T> {
     ///
     pub fn get_root_netuid() -> u16 {
         0
-    }
-
-    /// Fetches the total count of subnets.
-    ///
-    /// This function retrieves the total number of subnets present on the chain.
-    ///
-    /// # Returns:
-    /// * 'u16': The total number of subnets.
-    ///
-    pub fn get_num_subnets() -> u16 {
-        TotalNetworks::<T>::get()
-    }
-
-    /// Fetches the max number of subnet
-    ///
-    /// This function retrieves the max number of subnet.
-    ///
-    /// # Returns:
-    /// * 'u16': The max number of subnet
-    ///
-    pub fn get_max_subnets() -> u16 {
-        SubnetLimit::<T>::get()
-    }
-
-    /// Sets the max number of subnet
-    ///
-    /// This function sets the max number of subnet.
-    ///
-    pub fn set_max_subnets(limit: u16) {
-        SubnetLimit::<T>::put(limit);
-        Self::deposit_event(Event::SubnetLimitSet(limit));
     }
 
     /// Fetches the total count of root network validators
@@ -93,106 +58,6 @@ impl<T: Config> Pallet<T> {
         Self::get_max_allowed_uids(Self::get_root_netuid())
     }
 
-    /// Returns the emission value for the given subnet.
-    ///
-    /// This function retrieves the emission value for the given subnet.
-    ///
-    /// # Returns:
-    /// * 'u64': The emission value for the given subnet.
-    ///
-    pub fn get_subnet_emission_value(netuid: u16) -> u64 {
-        EmissionValues::<T>::get(netuid)
-    }
-
-    /// Returns true if the subnetwork exists.
-    ///
-    /// This function checks if a subnetwork with the given UID exists.
-    ///
-    /// # Returns:
-    /// * 'bool': Whether the subnet exists.
-    ///
-    pub fn if_subnet_exist(netuid: u16) -> bool {
-        NetworksAdded::<T>::get(netuid)
-    }
-
-    /// Returns a list of subnet netuid equal to total networks.
-    ///
-    ///
-    /// This iterates through all the networks and returns a list of netuids.
-    ///
-    /// # Returns:
-    /// * 'Vec<u16>': Netuids of all subnets.
-    ///
-    pub fn get_all_subnet_netuids() -> Vec<u16> {
-        <NetworksAdded<T> as IterableStorageMap<u16, bool>>::iter()
-            .map(|(netuid, _)| netuid)
-            .collect()
-    }
-
-    /// Calculates the block emission based on the total issuance.
-    ///
-    /// This function computes the block emission by applying a logarithmic function
-    /// to the total issuance of the network. The formula used takes into account
-    /// the current total issuance and adjusts the emission rate accordingly to ensure
-    /// a smooth issuance curve. The emission rate decreases as the total issuance increases,
-    /// following a logarithmic decay.
-    ///
-    /// # Returns
-    /// * 'Result<u64, &'static str>': The calculated block emission rate or error.
-    ///
-    pub fn get_block_emission() -> Result<u64, &'static str> {
-        // Convert the total issuance to a fixed-point number for calculation.
-        Self::get_block_emission_for_issuance(Self::get_total_issuance())
-    }
-
-    /// Returns the block emission for an issuance value.
-    pub fn get_block_emission_for_issuance(issuance: u64) -> Result<u64, &'static str> {
-        // Convert issuance to a float for calculations below.
-        let total_issuance: I96F32 = I96F32::from_num(issuance);
-        // Check to prevent division by zero when the total supply is reached
-        // and creating an issuance greater than the total supply.
-        if total_issuance >= I96F32::from_num(TotalSupply::<T>::get()) {
-            return Ok(0);
-        }
-        // Calculate the logarithmic residual of the issuance against half the total supply.
-        let residual: I96F32 = log2(
-            I96F32::from_num(1.0)
-                .checked_div(
-                    I96F32::from_num(1.0)
-                        .checked_sub(
-                            total_issuance
-                                .checked_div(
-                                    I96F32::from_num(2.0)
-                                        .saturating_mul(I96F32::from_num(10_500_000_000_000_000.0)),
-                                )
-                                .ok_or("Logarithm calculation failed")?,
-                        )
-                        .ok_or("Logarithm calculation failed")?,
-                )
-                .ok_or("Logarithm calculation failed")?,
-        )
-        .map_err(|_| "Logarithm calculation failed")?;
-        // Floor the residual to smooth out the emission rate.
-        let floored_residual: I96F32 = residual.floor();
-        // Calculate the final emission rate using the floored residual.
-        // Convert floored_residual to an integer
-        let floored_residual_int: u64 = floored_residual.to_num::<u64>();
-        // Multiply 2.0 by itself floored_residual times to calculate the power of 2.
-        let mut multiplier: I96F32 = I96F32::from_num(1.0);
-        for _ in 0..floored_residual_int {
-            multiplier = multiplier.saturating_mul(I96F32::from_num(2.0));
-        }
-        let block_emission_percentage: I96F32 = I96F32::from_num(1.0).saturating_div(multiplier);
-        // Calculate the actual emission based on the emission rate
-        let block_emission: I96F32 = block_emission_percentage
-            .saturating_mul(I96F32::from_num(DefaultBlockEmission::<T>::get()));
-        // Convert to u64
-        let block_emission_u64: u64 = block_emission.to_num::<u64>();
-        if BlockEmission::<T>::get() != block_emission_u64 {
-            BlockEmission::<T>::put(block_emission_u64);
-        }
-        Ok(block_emission_u64)
-    }
 
     /// Checks for any UIDs in the given list that are either equal to the root netuid or exceed the total number of subnets.
     ///
@@ -891,489 +756,5 @@ impl<T: Config> Pallet<T> {
             },
         )
             .into())
-    }
-
-    /// Facilitates user registration of a new subnetwork.
-    ///
-    /// # Args:
-    /// * 'origin': ('T::RuntimeOrigin'): The calling origin. Must be signed.
-    ///
-    /// # Event:
-    /// * 'NetworkAdded': Emitted when a new network is successfully added.
-    ///
-    /// # Raises:
-    /// * 'TxRateLimitExceeded': If the rate limit for network registration is exceeded.
-    /// * 'NotEnoughBalanceToStake': If there isn't enough balance to stake for network registration.
-    /// * 'BalanceWithdrawalError': If an error occurs during balance withdrawal for network registration.
-    ///
-    pub fn user_add_network(origin: T::RuntimeOrigin, hotkey: &T::AccountId, mechid: u16) -> dispatch::DispatchResult {
-        // --- 0. Ensure the caller is a signed user.
-        let coldkey = ensure_signed(origin)?;
-
-        // --- 1. Rate limit for network registrations.
-        let current_block = Self::get_current_block_as_u64();
-        let last_lock_block = Self::get_network_last_lock_block();
-        ensure!(
-            current_block.saturating_sub(last_lock_block) >= NetworkRateLimit::<T>::get(),
-            Error::<T>::NetworkTxRateLimitExceeded
-        );
-
-        // --- 2. Calculate and lock the required tokens.
-        let lock_amount: u64 = Self::get_network_lock_cost();
-        log::debug!("network lock_amount: {:?}", lock_amount);
-        ensure!(
-            Self::can_remove_balance_from_coldkey_account(&coldkey, lock_amount),
-            Error::<T>::NotEnoughBalanceToStake
-        );
-
-        // --- 4. Determine the netuid to register.
-        let netuid_to_register: u16 = {
-            log::debug!(
-                "subnet count: {:?}\nmax subnets: {:?}",
-                Self::get_num_subnets(),
-                Self::get_max_subnets()
-            );
-            if Self::get_num_subnets().saturating_sub(1) < Self::get_max_subnets() {
-                // We subtract one because we don't want root subnet to count towards total
-                let mut next_available_netuid = 0;
-                loop {
-                    next_available_netuid.saturating_inc();
-                    if !Self::if_subnet_exist(next_available_netuid) {
-                        log::debug!("got subnet id: {:?}", next_available_netuid);
-                        break next_available_netuid;
-                    }
-                }
-            } else {
-                let netuid_to_prune = Self::get_subnet_to_prune();
-                ensure!(netuid_to_prune > 0, Error::<T>::AllNetworksInImmunity);
-
-                Self::remove_network(netuid_to_prune);
-                log::debug!("remove_network: {:?}", netuid_to_prune,);
-                Self::deposit_event(Event::NetworkRemoved(netuid_to_prune));
-                netuid_to_prune
-            }
-        };
-
-        // --- 5. Perform the lock operation.
-        let actual_lock_amount = Self::remove_balance_from_coldkey_account(&coldkey, lock_amount)?;
-        log::debug!("actual_lock_amount: {:?}", actual_lock_amount);
-
-        // Self::set_subnet_locked_balance(netuid_to_register, actual_lock_amount);
-        Self::set_network_last_lock(actual_lock_amount);
-
-        // --- 6. Set initial and custom parameters for the network.
-        Self::init_new_network(netuid_to_register, 360);
-        log::debug!("init_new_network: {:?}", netuid_to_register);
-
-        // --- 7. Set netuid storage.
-        let current_block_number: u64 = Self::get_current_block_as_u64();
-        log::debug!("Current block number: {:?}", current_block_number);
-
-        NetworkLastRegistered::<T>::set(current_block_number);
-        log::debug!("NetworkLastRegistered set to: {:?}", current_block_number);
-
-        NetworkRegisteredAt::<T>::insert(netuid_to_register, current_block_number);
-        log::debug!("NetworkRegisteredAt for netuid {:?} set to: {:?}", netuid_to_register, current_block_number);
-
-        SubnetOwner::<T>::insert(netuid_to_register, coldkey.clone());
-        log::debug!("SubnetOwner for netuid {:?} set to: {:?}", netuid_to_register, coldkey);
-
-        SubnetMechanism::<T>::insert( netuid_to_register, mechid );
-        log::debug!("SubnetMechanism for netuid {:?} set to: {:?}", netuid_to_register, mechid);
-
-        Self::append_neuron( netuid_to_register, hotkey, current_block_number );
-        log::debug!("Appended neuron for netuid {:?}, hotkey: {:?}", netuid_to_register, hotkey);
-
-        // Compute the stake operation based on the mechanism.
-        let mechid: u16 = SubnetMechanism::<T>::get( netuid_to_register );
-        log::debug!("Mechanism ID: {:?}", mechid);
-
-        let alpha_amount_staked: u64;
-        if mechid == 2 { // STAO
-            // Compute dynamic stake.
-            let total_subnet_tao: u64 = SubnetTAO::<T>::get( netuid_to_register );
-            log::debug!("Total subnet TAO: {:?}", total_subnet_tao);
-
-            let total_mechanism_tao: u64 = Self::get_total_mechanism_tao( SubnetMechanism::<T>::get( netuid_to_register ) );
-            log::debug!("Total mechanism TAO: {:?}", total_mechanism_tao);
-
-            let price: I96F32 = I96F32::from_num(total_mechanism_tao + actual_lock_amount).checked_div(I96F32::from_num(total_subnet_tao + actual_lock_amount)).unwrap_or(I96F32::from_num(1));
-            log::debug!("price: {:?}", price);
-
-            alpha_amount_staked = (I96F32::from_num(actual_lock_amount) * price).to_num::<u64>();
-            log::debug!("Computed alpha amount staked (STAO): {:?}", alpha_amount_staked);
-        } else { // ROOT and other.
-            alpha_amount_staked = actual_lock_amount;
-            log::debug!("Alpha amount staked (ROOT/other): {:?}", alpha_amount_staked);
-        }
-
-        // Increment counters.
-        let new_subnet_alpha = SubnetAlpha::<T>::get(netuid_to_register).saturating_add(alpha_amount_staked);
-        SubnetAlpha::<T>::insert(netuid_to_register, new_subnet_alpha);
-        log::debug!("Updated SubnetAlpha for netuid {:?}: {:?}", netuid_to_register, new_subnet_alpha);
-
-        let new_subnet_tao = SubnetTAO::<T>::get(netuid_to_register).saturating_add( actual_lock_amount );
-        SubnetTAO::<T>::insert(netuid_to_register, actual_lock_amount);
-        log::debug!("Updated SubnetTAO for netuid {:?}: {:?} vs lock {:?}", netuid_to_register, new_subnet_tao, actual_lock_amount);
-
-        let new_stake = Stake::<T>::get( &hotkey, &coldkey ).saturating_add( actual_lock_amount );
-        Stake::<T>::insert(hotkey, &coldkey, new_stake);
-        log::debug!("Updated Stake for hotkey {:?}, coldkey {:?}: {:?}", hotkey, coldkey, new_stake);
-
-        let new_total_hotkey_alpha = TotalHotkeyAlpha::<T>::get( &hotkey, netuid_to_register ).saturating_add( alpha_amount_staked );
-        TotalHotkeyAlpha::<T>::insert(&hotkey, netuid_to_register, new_total_hotkey_alpha);
-        log::debug!("Updated TotalHotkeyAlpha for hotkey {:?}, netuid {:?}: {:?}", hotkey, netuid_to_register, new_total_hotkey_alpha);
-
-        let new_alpha = Alpha::<T>::get((&hotkey, &coldkey, netuid_to_register)).saturating_add( alpha_amount_staked );
-        Alpha::<T>::insert((&hotkey, &coldkey, netuid_to_register), new_alpha);
-        log::debug!("Updated Alpha for hotkey {:?}, coldkey {:?}, netuid {:?}: {:?}", hotkey, coldkey, netuid_to_register, new_alpha);
-
-        // --- 8. Emit the NetworkAdded event.
-        log::info!(
-            "NetworkAdded( netuid:{:?}, modality:{:?} )",
-            netuid_to_register,
-            0
-        );
-        Self::deposit_event(Event::NetworkAdded(netuid_to_register, 0));
-
-        // --- 9. Return success.
-        Ok(())
-    }
-
-    /// Facilitates the removal of a user's subnetwork.
-    ///
-    /// # Args:
-    /// * 'origin': ('T::RuntimeOrigin'): The calling origin. Must be signed.
-    /// * 'netuid': ('u16'): The unique identifier of the network to be removed.
-    ///
-    /// # Event:
-    /// * 'NetworkRemoved': Emitted when a network is successfully removed.
-    ///
-    /// # Raises:
-    /// * 'SubNetworkDoesNotExist': If the specified network does not exist.
-    /// * 'NotSubnetOwner': If the caller does not own the specified subnet.
-    ///
-    pub fn user_remove_network(origin: T::RuntimeOrigin, netuid: u16) -> dispatch::DispatchResult {
-        // --- 1. Ensure the function caller is a signed user.
-        let coldkey = ensure_signed(origin)?;
-
-        // --- 2. Ensure this subnet exists.
-        ensure!(
-            Self::if_subnet_exist(netuid),
-            Error::<T>::SubNetworkDoesNotExist
-        );
-
-        // --- 3. Ensure the caller owns this subnet.
-        ensure!(
-            SubnetOwner::<T>::get(netuid) == coldkey,
-            Error::<T>::NotSubnetOwner
-        );
-
-        // --- 4. Explicitly erase the network and all its parameters.
-        Self::remove_network(netuid);
-
-        // --- 5. Emit the NetworkRemoved event.
-        log::info!("NetworkRemoved( netuid:{:?} )", netuid);
-        Self::deposit_event(Event::NetworkRemoved(netuid));
-
-        // --- 6. Return success.
-        Ok(())
-    }
-
-    /// Sets initial and custom parameters for a new network.
-    pub fn init_new_network(netuid: u16, tempo: u16) {
-        // --- 1. Set network to 0 size.
-        SubnetworkN::<T>::insert(netuid, 0);
-
-        // --- 2. Set this network uid to alive.
-        NetworksAdded::<T>::insert(netuid, true);
-
-        // --- 3. Fill tempo memory item.
-        Tempo::<T>::insert(netuid, tempo);
-
-        // --- 4 Fill modality item.
-        NetworkModality::<T>::insert(netuid, 0);
-
-        // --- 5. Increase total network count.
-        TotalNetworks::<T>::mutate(|n| *n = n.saturating_add(1));
-
-        // --- 6. Set all default values **explicitly**.
-        Self::set_network_registration_allowed(netuid, true);
-        Self::set_max_allowed_uids(netuid, 256);
-        Self::set_max_allowed_validators(netuid, 64);
-        Self::set_min_allowed_weights(netuid, 1);
-        Self::set_max_weight_limit(netuid, u16::MAX);
-        Self::set_adjustment_interval(netuid, 360);
-        Self::set_target_registrations_per_interval(netuid, 1);
-        Self::set_adjustment_alpha(netuid, 17_893_341_751_498_265_066); // 18_446_744_073_709_551_615 * 0.97 = 17_893_341_751_498_265_066
-        Self::set_immunity_period(netuid, 5000);
-        Self::set_min_burn(netuid, 1);
-        Self::set_min_difficulty(netuid, u64::MAX);
-        Self::set_max_difficulty(netuid, u64::MAX);
-
-        // Make network parameters explicit.
-        if !Tempo::<T>::contains_key(netuid) {
-            Tempo::<T>::insert(netuid, Tempo::<T>::get(netuid));
-        }
-        if !Kappa::<T>::contains_key(netuid) {
-            Kappa::<T>::insert(netuid, Kappa::<T>::get(netuid));
-        }
-        if !Difficulty::<T>::contains_key(netuid) {
-            Difficulty::<T>::insert(netuid, Difficulty::<T>::get(netuid));
-        }
-        if !MaxAllowedUids::<T>::contains_key(netuid) {
-            MaxAllowedUids::<T>::insert(netuid, MaxAllowedUids::<T>::get(netuid));
-        }
-        if !ImmunityPeriod::<T>::contains_key(netuid) {
-            ImmunityPeriod::<T>::insert(netuid, ImmunityPeriod::<T>::get(netuid));
-        }
-        if !ActivityCutoff::<T>::contains_key(netuid) {
-            ActivityCutoff::<T>::insert(netuid, ActivityCutoff::<T>::get(netuid));
-        }
-        if !EmissionValues::<T>::contains_key(netuid) {
-            EmissionValues::<T>::insert(netuid, EmissionValues::<T>::get(netuid));
-        }
-        if !MaxWeightsLimit::<T>::contains_key(netuid) {
-            MaxWeightsLimit::<T>::insert(netuid, MaxWeightsLimit::<T>::get(netuid));
-        }
-        if !MinAllowedWeights::<T>::contains_key(netuid) {
-            MinAllowedWeights::<T>::insert(netuid, MinAllowedWeights::<T>::get(netuid));
-        }
-        if !RegistrationsThisInterval::<T>::contains_key(netuid) {
-            RegistrationsThisInterval::<T>::insert(
-                netuid,
-                RegistrationsThisInterval::<T>::get(netuid),
-            );
-        }
-        if !POWRegistrationsThisInterval::<T>::contains_key(netuid) {
-            POWRegistrationsThisInterval::<T>::insert(
-                netuid,
-                POWRegistrationsThisInterval::<T>::get(netuid),
-            );
-        }
-        if !BurnRegistrationsThisInterval::<T>::contains_key(netuid) {
-            BurnRegistrationsThisInterval::<T>::insert(
-                netuid,
-                BurnRegistrationsThisInterval::<T>::get(netuid),
-            );
-        }
-    }
-
-    /// Removes a network (identified by netuid) and all associated parameters.
-    ///
-    /// This function is responsible for cleaning up all the data associated with a network.
-    /// It ensures that all the storage values related to the network are removed, and any
-    /// reserved balance is returned to the network owner.
-    ///
-    /// # Args:
-    ///  * 'netuid': ('u16'): The unique identifier of the network to be removed.
-    ///
-    /// # Note:
-    /// This function does not emit any events, nor does it raise any errors. It silently
-    /// returns if any internal checks fail.
-    ///
-    pub fn remove_network(netuid: u16) {
-        // --- 1. Return balance to subnet owner.
-        let owner_coldkey = SubnetOwner::<T>::get(netuid);
-        let reserved_amount = Self::get_subnet_locked_balance(netuid);
-
-        // --- 2. Remove network count.
-        SubnetworkN::<T>::remove(netuid);
-
-        // --- 3. Remove network modality storage.
-        NetworkModality::<T>::remove(netuid);
-
-        // --- 4. Remove netuid from added networks.
-        NetworksAdded::<T>::remove(netuid);
-
-        // --- 6. Decrement the network counter.
-        TotalNetworks::<T>::mutate(|n| *n = n.saturating_sub(1));
-
-        // --- 7. Remove various network-related storages.
-        NetworkRegisteredAt::<T>::remove(netuid);
-
-        // --- 8. Remove incentive mechanism memory.
-        let _ = Uids::<T>::clear_prefix(netuid, u32::MAX, None);
-        let _ = Keys::<T>::clear_prefix(netuid, u32::MAX, None);
-        let _ = Bonds::<T>::clear_prefix(netuid, u32::MAX, None);
-
-        // --- 8. Removes the weights for this subnet (do not remove).
-        let _ = Weights::<T>::clear_prefix(netuid, u32::MAX, None);
-
-        // --- 9. Iterate over stored weights and fill the matrix.
-        for (uid_i, weights_i) in
-            <Weights<T> as IterableStorageDoubleMap<u16, u16, Vec<(u16, u16)>>>::iter_prefix(
-                Self::get_root_netuid(),
-            )
-        {
-            // Create a new vector to hold modified weights.
-            let mut modified_weights = weights_i.clone();
-            // Iterate over each weight entry to potentially update it.
-            for (subnet_id, weight) in modified_weights.iter_mut() {
-                if subnet_id == &netuid {
-                    // If the condition matches, modify the weight
-                    *weight = 0; // Set weight to 0 for the matching subnet_id.
-                }
-            }
-            Weights::<T>::insert(Self::get_root_netuid(), uid_i, modified_weights);
-        }
-
-        // --- 10. Remove various network-related parameters.
-        Rank::<T>::remove(netuid);
-        Trust::<T>::remove(netuid);
-        Active::<T>::remove(netuid);
-        Emission::<T>::remove(netuid);
-        Incentive::<T>::remove(netuid);
-        Consensus::<T>::remove(netuid);
-        Dividends::<T>::remove(netuid);
-        PruningScores::<T>::remove(netuid);
-        LastUpdate::<T>::remove(netuid);
-        ValidatorPermit::<T>::remove(netuid);
-        ValidatorTrust::<T>::remove(netuid);
-
-        // --- 11. Erase network parameters.
-        Tempo::<T>::remove(netuid);
-        Kappa::<T>::remove(netuid);
-        Difficulty::<T>::remove(netuid);
-        MaxAllowedUids::<T>::remove(netuid);
-        ImmunityPeriod::<T>::remove(netuid);
-        ActivityCutoff::<T>::remove(netuid);
-        EmissionValues::<T>::remove(netuid);
-        MaxWeightsLimit::<T>::remove(netuid);
-        MinAllowedWeights::<T>::remove(netuid);
-        RegistrationsThisInterval::<T>::remove(netuid);
-        POWRegistrationsThisInterval::<T>::remove(netuid);
-        BurnRegistrationsThisInterval::<T>::remove(netuid);
-
-        // --- 12. Add the balance back to the owner.
-        Self::add_balance_to_coldkey_account(&owner_coldkey, reserved_amount);
-        Self::set_subnet_locked_balance(netuid, 0);
-        SubnetOwner::<T>::remove(netuid);
-    }
-
-    #[allow(clippy::arithmetic_side_effects)]
-    /// This function calculates the lock cost for a network based on the last lock amount, minimum lock cost, last lock block, and current block.
-    /// The lock cost is calculated using the formula:
-    /// lock_cost = (last_lock * mult) - (last_lock / lock_reduction_interval) * (current_block - last_lock_block)
-    /// where:
-    /// - last_lock is the last lock amount for the network
-    /// - mult is the multiplier which increases lock cost each time a registration occurs
-    /// - last_lock_block is the block number at which the last lock occurred
-    /// - lock_reduction_interval the number of blocks before the lock returns to previous value.
-    /// - current_block is the current block number
-    /// - DAYS is the number of blocks in a day
-    /// - min_lock is the minimum lock cost for the network
-    ///
-    /// If the calculated lock cost is less than the minimum lock cost, the minimum lock cost is returned.
-    ///
-    /// # Returns:
-    ///  * 'u64':
-    ///     - The lock cost for the network.
-    ///
-    pub fn get_network_lock_cost() -> u64 {
-        let last_lock = Self::get_network_last_lock();
-        let min_lock = Self::get_network_min_lock();
-        let last_lock_block = Self::get_network_last_lock_block();
-        let current_block = Self::get_current_block_as_u64();
-        let lock_reduction_interval = Self::get_lock_reduction_interval();
-        let mult = if last_lock_block == 0 { 1 } else { 2 };
-
-        let mut lock_cost = last_lock.saturating_mul(mult).saturating_sub(
-            last_lock
-                .saturating_div(lock_reduction_interval)
-                .saturating_mul(current_block.saturating_sub(last_lock_block)),
-        );
-
-        if lock_cost < min_lock {
-            lock_cost = min_lock;
-        }
-
-        log::debug!( "last_lock: {:?}, min_lock: {:?}, last_lock_block: {:?}, lock_reduction_interval: {:?}, current_block: {:?}, mult: {:?} lock_cost: {:?}",
-        last_lock, min_lock, last_lock_block, lock_reduction_interval, current_block, mult, lock_cost);
-
-        lock_cost
-    }
-
-    /// This function is used to determine which subnet to prune when the total number of networks has reached the limit.
-    /// It iterates over all the networks and finds the oldest subnet with the minimum emission value that is not in the immunity period.
-    ///
-    /// # Returns:
-    /// * 'u16':
-    ///     - The uid of the network to be pruned.
-    ///
-    pub fn get_subnet_to_prune() -> u16 {
-        let mut netuids: Vec<u16> = vec![];
-        let current_block = Self::get_current_block_as_u64();
-
-        // Even if we don't have a root subnet, this still works
-        for netuid in NetworksAdded::<T>::iter_keys_from(NetworksAdded::<T>::hashed_key_for(0)) {
-            if current_block.saturating_sub(Self::get_network_registered_block(netuid))
-                < Self::get_network_immunity_period()
-            {
-                continue;
-            }
-
-            // This iterator seems to return them in order anyways, so no need to sort by key
-            netuids.push(netuid);
-        }
-
-        // Now we sort by emission, and then by subnet creation time.
-        netuids.sort_by(|a, b| {
-            use sp_std::cmp::Ordering;
-
-            match Self::get_emission_value(*b).cmp(&Self::get_emission_value(*a)) {
-                Ordering::Equal => {
-                    if Self::get_network_registered_block(*b)
-                        < Self::get_network_registered_block(*a)
-                    {
-                        Ordering::Less
-                    } else {
-                        Ordering::Equal
-                    }
-                }
-                v => v,
-            }
-        });
-
-        log::info!("Netuids Order: {:?}", netuids);
-
-        match netuids.last() {
-            Some(netuid) => *netuid,
-            None => 0,
-        }
-    }
-
-    pub fn get_network_registered_block(netuid: u16) -> u64 {
-        NetworkRegisteredAt::<T>::get(netuid)
-    }
-    pub fn get_network_immunity_period() -> u64 {
-        NetworkImmunityPeriod::<T>::get()
-    }
-    pub fn set_network_immunity_period(net_immunity_period: u64) {
-        NetworkImmunityPeriod::<T>::set(net_immunity_period);
-        Self::deposit_event(Event::NetworkImmunityPeriodSet(net_immunity_period));
-    }
-    pub fn set_network_min_lock(net_min_lock: u64) {
-        NetworkMinLockCost::<T>::set(net_min_lock);
-        Self::deposit_event(Event::NetworkMinLockCostSet(net_min_lock));
-    }
-    pub fn get_network_min_lock() -> u64 {
-        NetworkMinLockCost::<T>::get()
-    }
-    pub fn set_network_last_lock(net_last_lock: u64) {
-        NetworkLastLockCost::<T>::set(net_last_lock);
-    }
-    pub fn get_network_last_lock() -> u64 {
-        NetworkLastLockCost::<T>::get()
-    }
-    pub fn get_network_last_lock_block() -> u64 {
-        NetworkLastRegistered::<T>::get()
-    }
-    pub fn set_lock_reduction_interval(interval: u64) {
-        NetworkLockReductionInterval::<T>::set(interval);
-        Self::deposit_event(Event::NetworkLockCostReductionIntervalSet(interval));
-    }
-    pub fn get_lock_reduction_interval() -> u64 {
-        NetworkLockReductionInterval::<T>::get()
     }
 }
