+++ pallets/subtensor/src/subnets/lock.rs
@@ -1,0 +1,82 @@
+
+use super::*;
+
+impl<T: Config> Pallet<T> {
+    
+    #[allow(clippy::arithmetic_side_effects)]
+    /// This function calculates the lock cost for a network based on the last lock amount, minimum lock cost, last lock block, and current block.
+    /// The lock cost is calculated using the formula:
+    /// lock_cost = (last_lock * mult) - (last_lock / lock_reduction_interval) * (current_block - last_lock_block)
+    /// where:
+    /// - last_lock is the last lock amount for the network
+    /// - mult is the multiplier which increases lock cost each time a registration occurs
+    /// - last_lock_block is the block number at which the last lock occurred
+    /// - lock_reduction_interval the number of blocks before the lock returns to previous value.
+    /// - current_block is the current block number
+    /// - DAYS is the number of blocks in a day
+    /// - min_lock is the minimum lock cost for the network
+    ///
+    /// If the calculated lock cost is less than the minimum lock cost, the minimum lock cost is returned.
+    ///
+    /// # Returns:
+    ///  * 'u64':
+    ///     - The lock cost for the network.
+    ///
+    pub fn get_network_lock_cost() -> u64 {
+        let last_lock = Self::get_network_last_lock();
+        let min_lock = Self::get_network_min_lock();
+        let last_lock_block = Self::get_network_last_lock_block();
+        let current_block = Self::get_current_block_as_u64();
+        let lock_reduction_interval = Self::get_lock_reduction_interval();
+        let mult = if last_lock_block == 0 { 1 } else { 2 };
+
+        let mut lock_cost = last_lock.saturating_mul(mult).saturating_sub(
+            last_lock
+                .saturating_div(lock_reduction_interval)
+                .saturating_mul(current_block.saturating_sub(last_lock_block)),
+        );
+
+        if lock_cost < min_lock {
+            lock_cost = min_lock;
+        }
+
+        log::debug!( "last_lock: {:?}, min_lock: {:?}, last_lock_block: {:?}, lock_reduction_interval: {:?}, current_block: {:?}, mult: {:?} lock_cost: {:?}",
+        last_lock, min_lock, last_lock_block, lock_reduction_interval, current_block, mult, lock_cost);
+
+        lock_cost
+    }
+
+    pub fn get_network_registered_block(netuid: u16) -> u64 {
+        NetworkRegisteredAt::<T>::get(netuid)
+    }
+    pub fn get_network_immunity_period() -> u64 {
+        NetworkImmunityPeriod::<T>::get()
+    }
+    pub fn set_network_immunity_period(net_immunity_period: u64) {
+        NetworkImmunityPeriod::<T>::set(net_immunity_period);
+        Self::deposit_event(Event::NetworkImmunityPeriodSet(net_immunity_period));
+    }
+    pub fn set_network_min_lock(net_min_lock: u64) {
+        NetworkMinLockCost::<T>::set(net_min_lock);
+        Self::deposit_event(Event::NetworkMinLockCostSet(net_min_lock));
+    }
+    pub fn get_network_min_lock() -> u64 {
+        NetworkMinLockCost::<T>::get()
+    }
+    pub fn set_network_last_lock(net_last_lock: u64) {
+        NetworkLastLockCost::<T>::set(net_last_lock);
+    }
+    pub fn get_network_last_lock() -> u64 {
+        NetworkLastLockCost::<T>::get()
+    }
+    pub fn get_network_last_lock_block() -> u64 {
+        NetworkLastRegistered::<T>::get()
+    }
+    pub fn set_lock_reduction_interval(interval: u64) {
+        NetworkLockReductionInterval::<T>::set(interval);
+        Self::deposit_event(Event::NetworkLockCostReductionIntervalSet(interval));
+    }
+    pub fn get_lock_reduction_interval() -> u64 {
+        NetworkLockReductionInterval::<T>::get()
+    }
+}
