+++ pallets/subtensor/src/subnets/mod.rs
@@ -3,3 +3,6 @@ pub mod registration;
 pub mod serving;
 pub mod uids;
 pub mod weights;
+pub mod create;
+pub mod lock;
+
